+++ lib/index.js
@@ -5,10 +5,9 @@ const yaml = require('js-yaml')
 const Joi = require('joi')
 const hostname = require('@runnable/hostname')
 
-const removeNullAndUndefinedKeys = require('./util').removeNullAndUndefinedKeys
 const validate = require('./util').validate
-const Parser = require('./parser')
 const Mapper = require('./mapper')
+const Parser = require('./parser')
 const Warning = require('./warning')
 
 const dockerComposeFileSchema = Joi.object({
@@ -52,9 +51,9 @@ module.exports = class DockerComposeParser {
 
     return DockerComposeParser._parseDockerComposeFile(dockerComposeFileString)
     .then(services => DockerComposeParser._convertToArrayAndAddMetadata(services))
-    .map(service => DockerComposeParser._serviceParser({ service, repositoryName }))
-    .map(service => DockerComposeParser._addHostname(service, ownerUsername, userContentDomain))
-    .then(services => DockerComposeParser._populateEnvsWithHost(services))
+    .map(service => DockerComposeParser._parseService({ service, repositoryName }))
+    .map(service => DockerComposeParser._populateHostname(service, ownerUsername, userContentDomain))
+    .then(services => DockerComposeParser._populateEnvsWithHostFromServices(services))
     .then(services => { return { results: services } })
   }
 
@@ -98,14 +97,14 @@ module.exports = class DockerComposeParser {
   }
 
   /**
-   * Adds a hostname to a service
+   * Populates the hostname for a service
    *
    * @param {Object} service
    * @param {Object} service.metadata      - Metadata object
    * @param {String} service.instance.name - Name of the new instance
    * @return {Object} - Returns same object
    */
-  static _addHostname (service, ownerUsername, userContentDomain) {
+  static _populateHostname (service, ownerUsername, userContentDomain) {
     service.metadata.hostname = hostname.elastic({
       shortHash: '000', // This is required, but ignored
       instanceName: service.instance.name,
@@ -125,7 +124,7 @@ module.exports = class DockerComposeParser {
    * @param {Array<String>} services.instance.env - ENVs for new instance
    * @return {Array<Service>} - Returns array with transformed objects
    */
-  static _populateEnvsWithHost (services) {
+  static _populateEnvsWithHostFromServices (services) {
     const hostnames = services.reduce((obj, service) => {
       obj[service.metadata.name] = service.metadata.hostname
       return obj
@@ -164,7 +163,7 @@ module.exports = class DockerComposeParser {
    * @return {Promise} - Returns Object with service
    * @resolves {Object}
    */
-  static _serviceParser (opts) {
+  static _parseService (opts) {
     const service = opts.service
     const repositoryName = opts.repositoryName
 
@@ -181,11 +180,11 @@ module.exports = class DockerComposeParser {
     .then(rawService => Warning.unsupportedKeys(rawService, warnings, serviceSchemaKeys))
     .then(rawService => Warning.buildAndImage(rawService, warnings))
     // Populate properties
-    .then(rawService => DockerComposeParser._mapRaseServiceToService(service, rawService, repositoryName, warnings))
+    .then(rawService => DockerComposeParser._mapRawServiceToService(service, rawService, repositoryName, warnings))
     .then(service => DockerComposeParser._removeUnusedKeys(service, warnings))
   }
 
-  static _mapRaseServiceToService (service, rawService, repositoryName, warnings) {
+  static _mapRawServiceToService (service, rawService, repositoryName, warnings) {
     return Promise.props({
       metadata: service.metadata,
       contextVersion: Promise.props({
@@ -211,10 +210,10 @@ module.exports = class DockerComposeParser {
   static _removeUnusedKeys (service, warnings) {
     Object.assign(service, {
       warnings: Array.from(warnings), // Convert to array from iterator
-      contextVersion: removeNullAndUndefinedKeys(service.contextVersion),
-      instance: removeNullAndUndefinedKeys(service.instance)
+      contextVersion: Mapper.removeNullAndUndefinedKeys(service.contextVersion),
+      instance: Mapper.removeNullAndUndefinedKeys(service.instance)
     })
-    return removeNullAndUndefinedKeys(service)
+    return Mapper.removeNullAndUndefinedKeys(service)
   }
 
 }
