+++ lib/parser.js
@@ -35,7 +35,6 @@ module.exports = class Parser {
     if (!build) return null
 
     if (build.args) {
-      // NOTE: Maybe pass these to the ENVs?
       warnings.add('The `args` argument is not supported for builds. These args will not be passed to the build.', { args: build.args })
     }
