+++ lib/util.js
@@ -4,23 +4,6 @@ const Promise = require('bluebird')
 module.exports = class Utils {
 
   /**
-   * Remove all null and undefined values in object. Shallow check.
-   *
-   * @param {Object}
-   * @return {Object} - Shallow of object without keys/values
-   */
-  static removeNullAndUndefinedKeys (obj) {
-    const newObject = Object.assign({}, obj)
-    Object.keys(newObject).forEach(key => {
-      const value = obj[key]
-      if (value === undefined || value === null) {
-        delete newObject[key]
-      }
-    })
-    return newObject
-  }
-
-  /**
    * Validate payload against Joi schema
    *
    * @param {Object}   schema  - Joi schema
